requirements.txt
amulet_core==1.9.27
